da controllare hitbox entity

+++ src/view/entityView/EntityView.java
@@ -115,8 +115,8 @@ public abstract class EntityView extends SortableElement {
 
     protected void getCurrentDirectionFromController() {
 
-        float vectorX = view.getModel().getCurrentDirectionOfEntity(indexInEntityArray).getX();
-        float vectorY = view.getModel().getCurrentDirectionOfEntity(indexInEntityArray).getY();
+        int vectorX = view.getModel().getCurrentDirectionOfEntity(indexInEntityArray).getX();
+        int vectorY = view.getModel().getCurrentDirectionOfEntity(indexInEntityArray).getY();
 
         if(vectorY < 0) {
             currentDirection = UP;
